Update ILC communication test. Add bump test.
